Исправил ошибки в модуляции
