+++ index.html
@@ -11,7 +11,6 @@
 
 <body>
   <main>
-    <script src="js/function.js"></script>
     <script type="module" src="js/main.js"></script>
     <!-- Фильтрация изображений от других пользователей -->
     <section class="img-filters  img-filters--inactive  container">
