+++ js/main.js
@@ -1,2 +1,3 @@
-import './data.js';
+import { getPhotoCards } from './create-photo-cards.js';
 
+console.log(getPhotoCards());
